+++ barsik_bot/__init__.py
@@ -1,0 +1,14 @@
+__author__ = "TimyrTM"
+__version__ = "0.4"
+
+from barsik_bot.bot import BarsikBot
+from barsik_bot.locale import BotLocalizationStore
+from barsik_bot.config import Config
+from barsik_bot.docs import (
+    Docs,
+    SlashCommandOptionDocs,
+    SlashCommandDocs,
+    DockLoader,
+    DocsNotFound,
+)
+from barsik_bot.errors import ErrorEmbed, ErrorHandler