+++ barsik_bot/utils.py
@@ -1,0 +1,5 @@
+def list_format(items: list[str]):
+    data = ""
+    for item in items:
+        data += f"- {item}\n"
+    return data