+++ changelog.md
@@ -1,6 +1,17 @@
-# Changelog 
+# 😼 Changelog
+
 ___
+### Version 0.4 ⭐
+- updated command: `/help`
+- updated bot documentation system
+- updated bot error system
+- updated bot localization system
+
+### Version 0.3
+- remove: `/clear [count] {channel} {member}`
+
 ### Version 0.2
 - add command: `/clear [count] {channel} {member}`
+
 ### Version 0.1
 - add command: `/help [command]`